new helper file to refactor all sorting and search functions to
